Move ScrollComponent to kas_core::event::components
