+++ crates/kas-widgets/src/lib.rs
@@ -98,7 +98,7 @@ pub use menu::*;
 pub use nav_frame::NavFrame;
 pub use progress::ProgressBar;
 pub use radiobox::{RadioBox, RadioBoxBare, RadioBoxGroup};
-pub use scroll::{ScrollComponent, ScrollRegion};
+pub use scroll::ScrollRegion;
 pub use scroll_label::ScrollLabel;
 pub use scrollbar::{ScrollBar, ScrollBarRegion, ScrollBars, Scrollable};
 pub use separator::Separator;
